+++ app/src/main/AndroidManifest.xml
@@ -20,9 +20,6 @@
         android:theme="@style/Theme.GodFisherman"
         android:usesCleartextTraffic="true">
         <activity
-            android:name=".ui.stopwatch.StopwatchActivity2"
-            android:exported="true"/>
-        <activity
             android:name=".ui.camera.upload.UploadActivity"
             android:exported="false" />
         <activity
