--- app/src/main/java/com/android04/godfisherman/ui/stopwatch/StopWatchInfoTestViewModel.kt
@@ -1,13 +1,0 @@
-package com.android04.godfisherman.ui.stopwatch
-
-import androidx.lifecycle.ViewModel
-import com.android04.godfisherman.data.repository.StopwatchRepository
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class StopWatchInfoTestViewModel  @Inject constructor(
-    private val repository: StopwatchRepository
-) : ViewModel() {
-
-}
