+++ app/src/main/java/com/android04/godfisherman/ui/stopwatch/StopwatchViewModel.kt
@@ -3,17 +3,27 @@ package com.android04.godfisherman.ui.stopwatch
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import com.android04.godfisherman.data.repository.StopwatchRepository
 import com.android04.godfisherman.utils.toTimeMilliSecond
+import dagger.hilt.android.lifecycle.HiltViewModel
 import java.lang.Thread.sleep
 import java.util.*
+import javax.inject.Inject
 
-class StopwatchViewModel: ViewModel() {
+@HiltViewModel
+class StopwatchViewModel @Inject constructor(
+    private val repository: StopwatchRepository
+): ViewModel() {
 
-    private lateinit var stopwatch: Timer
-    var time = 0.0
+    companion object{
+        var isTimeLine = false
+    }
 
     private val _isStopwatchStarted = MutableLiveData<Boolean>(false)
     val isStopwatchStarted: LiveData<Boolean> = _isStopwatchStarted
+
+    private lateinit var stopwatch: Timer
+    var time = 0.0
 
     private val _displayTime = MutableLiveData<String>("00:00:00.00")
     val displayTime: LiveData<String> = _displayTime
